requirements.txt
aiogram==3.15.0
aiohttp==3.9.5
asyncio-mqtt==0.16.1
pymysql==1.1.0
aiomysql==0.2.0
python-dotenv==1.0.0
qrcode==7.4.2
pillow==10.4.0
cryptography==42.0.5
redis==5.0.1
aioredis==2.0.1
requests==2.31.0
uuid==1.30
datetime
asyncio
logging
hashlib
secrets
random